Add PlaygroundActivity for demos stuff
- Moved demo stuff from MainActivity to PlaygroundActivity
