+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">lites-android</string>
+    <!-- Playground Activity -->
+    <string name="playground_activity">Playground Activity</string>
 </resources>
